refactor: Change the package structure for integrated management of reservation and use
