+++ src/main/java/project/seatsence/src/utilization/api/reservation/UserReservationApi.java
@@ -1,7 +1,7 @@
-package project.seatsence.src.utilization.api;
+package project.seatsence.src.utilization.api.reservation;
 
 import static project.seatsence.global.code.ResponseCode.*;
-import static project.seatsence.src.utilization.domain.ReservationStatus.*;
+import static project.seatsence.src.utilization.domain.reservation.ReservationStatus.*;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -15,12 +15,12 @@ import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import project.seatsence.global.exceptions.BaseException;
 import project.seatsence.global.response.SliceResponse;
-import project.seatsence.src.utilization.domain.Reservation;
-import project.seatsence.src.utilization.dto.request.SeatReservationRequest;
-import project.seatsence.src.utilization.dto.request.SpaceReservationRequest;
-import project.seatsence.src.utilization.dto.response.UserReservationListResponse;
-import project.seatsence.src.utilization.service.ReservationService;
-import project.seatsence.src.utilization.service.UserReservationService;
+import project.seatsence.src.utilization.domain.reservation.Reservation;
+import project.seatsence.src.utilization.dto.reservation.request.SeatReservationRequest;
+import project.seatsence.src.utilization.dto.reservation.request.SpaceReservationRequest;
+import project.seatsence.src.utilization.dto.reservation.response.UserReservationListResponse;
+import project.seatsence.src.utilization.service.reservation.ReservationService;
+import project.seatsence.src.utilization.service.reservation.UserReservationService;
 import project.seatsence.src.store.domain.Store;
 import project.seatsence.src.store.domain.StoreChair;
 import project.seatsence.src.store.domain.StoreSpace;
