+++ src/main/java/project/seatsence/src/utilization/service/reservation/AdminReservationService.java
@@ -1,4 +1,4 @@
-package project.seatsence.src.utilization.service;
+package project.seatsence.src.utilization.service.reservation;
 
 import static project.seatsence.global.code.ResponseCode.RESERVATION_NOT_FOUND;
 
@@ -7,9 +7,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import project.seatsence.global.exceptions.BaseException;
-import project.seatsence.src.utilization.dao.ReservationRepository;
-import project.seatsence.src.utilization.domain.Reservation;
-import project.seatsence.src.utilization.domain.ReservationStatus;
+import project.seatsence.src.utilization.dao.reservation.ReservationRepository;
+import project.seatsence.src.utilization.domain.reservation.Reservation;
+import project.seatsence.src.utilization.domain.reservation.ReservationStatus;
 
 @Service
 @Transactional
