+++ src/main/java/project/seatsence/src/utilization/service/reservation/UserReservationService.java
@@ -1,8 +1,8 @@
-package project.seatsence.src.utilization.service;
+package project.seatsence.src.utilization.service.reservation;
 
 import static project.seatsence.global.constants.Constants.*;
 import static project.seatsence.global.entity.BaseTimeAndStateEntity.State.*;
-import static project.seatsence.src.utilization.domain.ReservationStatus.*;
+import static project.seatsence.src.utilization.domain.reservation.ReservationStatus.*;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -11,10 +11,10 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import project.seatsence.global.response.SliceResponse;
-import project.seatsence.src.utilization.dao.ReservationRepository;
-import project.seatsence.src.utilization.domain.Reservation;
-import project.seatsence.src.utilization.domain.ReservationStatus;
-import project.seatsence.src.utilization.dto.response.UserReservationListResponse;
+import project.seatsence.src.utilization.dao.reservation.ReservationRepository;
+import project.seatsence.src.utilization.domain.reservation.Reservation;
+import project.seatsence.src.utilization.domain.reservation.ReservationStatus;
+import project.seatsence.src.utilization.dto.reservation.response.UserReservationListResponse;
 import project.seatsence.src.user.domain.User;
 import project.seatsence.src.user.service.UserService;
